requirements.txt
streamlit
pandas
numpy
bokeh
QuantStats
Pillow
TA-Lib == 0.4.25
